Timer fix
Ensure that timer compares are not set below current timer count, update timer count if they are. Otherwise, stepper ISR stops until 32-bit counter overflows. There's probably a better way to check or address this - appears to be the purpose of HAL_timer_restrain(), but still occurs even with that in place.